In actor favorite count graph, show data points that are not snapshotted
This has an issue where, after 'today' 6am JST but backend has not collected the data yet, we draw a wrong data point in the graph. But since frontend anyways has no way to know if backend has finished the job, it should be fine to keep it for now.