Update (some) dependencies
We should technically be able to migrate to go 1.12+ given that app_engine_apis would give us access to legacy memcache.

However, github.com/qedus/nds supports v2 memcache only in the v2 branch (which works with the new datastore API instead of appengine API). We may need to fork it to get it work with v2 appengine API first.